+++ src/renderer/components/message/MessageListAndComposer.tsx
@@ -128,11 +128,7 @@ export default function MessageListAndComposer({
       onDragOver={onDragOver}
     >
       <div className='message-list-and-composer__message-list'>
-        <MessageList
-          chat={chat}
-          refComposer={refComposer}
-          locationStreamingEnabled={settings.enableOnDemandLocationStreaming}
-        />
+        <MessageList chat={chat} refComposer={refComposer} />
       </div>
       <Composer
         ref={refComposer}