+++ src/shared/shared-types.d.ts
@@ -154,13 +154,22 @@ export interface MessageTypeAttachment {
   fileSize: string
 }
 
+export type msgStatus =
+  | 'error'
+  | 'sending'
+  | 'draft'
+  | 'delivered'
+  | 'read'
+  | 'sent'
+  | ''
+
 export interface MessageType {
   id: number
   msg: JsonMessage & {
     sentAt: number
     receivedAt: number
     direction: 'outgoing' | 'incoming'
-    status: todo
+    status: msgStatus
     attachment?: MessageTypeAttachment
   }
   filemime: string